XML handling of chars outside normal range fail
xml.c was a hang, xmlserializer.c was a crash.
Fixes issue #196

+++ Unix/xmlserializer/xmlserializer.c
@@ -890,7 +890,7 @@ static void WriteBuffer_Char(
     };
 
     if ((escapingDepth == 0) ||
-        (charToWrite > L'>') ||
+        (((unsigned int)charToWrite) > L'>') ||
         (_encode[(unsigned int)charToWrite].str == NULL))
     {
         WriteBuffer_NoneEscapedChar(clientBuffer, clientBufferLength, clientBufferNeeded, charToWrite, result);
